Added github stars
